requirements.txt
selenium
openai
pillow
requests
python-dotenv